Provisioning: further consolidate integration test packages
- Move export_folders_flag_disabled_test.go from root to jobs/
  (root now uses WithoutExportFeatureFlag)
- Move jobs/disabledfeatures/ tests to root provisioning/ package
  (compatible since root now disables export feature)
- Rename git/foldermetadata_incremental/ to foldermetadata/incremental/
- Update README with new package layout

Made-with: Cursor
