+++ pkg/tests/apis/provisioning/foldermetadata/incremental/helper_test.go
@@ -1,4 +1,4 @@
-package foldermetadataincremental
+package incremental
 
 import (
 	"testing"
