+++ pkg/tests/apis/provisioning/foldermetadata/incremental/incremental_invalid_folder_metadata_test.go
@@ -1,4 +1,4 @@
-package foldermetadataincremental
+package incremental
 
 import (
 	"context"
