+++ pkg/tests/apis/provisioning/foldermetadata/incremental/incremental_permissions_move_test.go
@@ -1,4 +1,4 @@
-package foldermetadataincremental
+package incremental
 
 import (
 	"context"
