+++ pkg/tests/apis/provisioning/jobs/export_folders_flag_disabled_test.go
@@ -1,4 +1,4 @@
-package provisioning
+package jobs
 
 import (
 	"os"
